+++ ConfluenceOtherworld/src/main/java/org/confluence/mod/integration/ars_nouveau/ArsNouveauHelper.java
@@ -32,7 +32,7 @@ public class ArsNouveauHelper {
     }
 
     public static void extractMana(ServerPlayer serverPlayer, ManaStorage manaStorage, double manaToRemove) {
-        PlayerUtils.extractAndDelayAndSync(manaStorage, () -> (int) (manaToRemove * toConfluence()), serverPlayer);
+        PlayerUtils.extractAndDelayAndSync(manaStorage, () -> (float) (manaToRemove * toConfluence()), serverPlayer);
     }
 
     public static TriState enoughMana(LivingEntity living, int totalCost) {
